+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v2/registry.ts
@@ -23,7 +23,7 @@ import { EnableTableEmptyRowModeQuickAction } from './op-enable-empty-row-mode';
 import { AddNewAnnotationFile } from '../common/add-new-annotation-file';
 import { EnableObjectPageVariantManagementQuickAction } from './op-enable-variant-management';
 import { EnableListReportVariantManagementQuickAction } from './lr-enable-variant-management';
-import { AddNewSubpage } from '../common/add-new-subpage';
+import { AddNewSubpage } from '../fe-v2/add-new-subpage';
 type PageName = 'listReport' | 'objectPage' | 'analyticalListPage';
 
 const OBJECT_PAGE_TYPE = 'sap.suite.ui.generic.template.ObjectPage.view.Details';