+++ packages/ui5-application-inquirer/test/unit/prompts/validators.test.ts
@@ -2,9 +2,7 @@ import * as projectValidators from '@sap-ux/project-input-validator';
 import * as promptHelpers from '../../../src/prompts/prompt-helpers';
 import { join } from 'path';
 import { initI18nUi5AppInquirer, t } from '../../../src/i18n';
-import { validateAppName, validateFioriAppProjectFolder } from '../../../src/prompts/validators';
-import { findRootsForPath, findCapProjectRoot } from '@sap-ux/project-access';
-import * as projectAccess from '@sap-ux/project-access';
+import { validateAppName } from '../../../src/prompts/validators';
 
 /**
  * Workaround to allow spyOn
@@ -49,38 +47,5 @@ describe('validators', () => {
         expect(validateAppName(appName, targetPath)).toBe(
             `A module with this name already exists in the folder: ${targetPath}`
         );
-    });
-    describe('validateFioriAppProjectFolder', () => {
-        const mockFindRootsForPath = jest.fn();
-        const mockFindCapProjectRoot = jest.fn();
-
-        beforeEach(() => {
-            (findRootsForPath as jest.Mock) = mockFindRootsForPath;
-            (findCapProjectRoot as jest.Mock) = mockFindCapProjectRoot;
-            jest.clearAllMocks();
-        });
-
-        test('should return true if no Fiori project is found in the target directory', async () => {
-            mockFindRootsForPath.mockResolvedValueOnce(null);
-            const result = await validateFioriAppProjectFolder('/path/to/dir');
-            expect(result).toBe(true);
-            expect(mockFindRootsForPath).toHaveBeenCalledWith('/path/to/dir');
-        });
-
-        test('should return an error message if a Fiori project is found in the target directory', async () => {
-            const appRootPath = '/path/to/fiori/project';
-            const projectRootPath = 'test/path';
-            mockFindRootsForPath.mockResolvedValueOnce({ appRoot: appRootPath, projectRoot: projectRootPath });
-            const result = await validateFioriAppProjectFolder('some/path');
-            expect(result).toEqual(t('validators.folderContainsFioriApp', { path: appRootPath }));
-            expect(mockFindRootsForPath).toHaveBeenCalledWith('some/path');
-        });
-
-        test('should return an error message if a CAP project is found in the target directory', async () => {
-            mockFindRootsForPath.mockResolvedValueOnce(null);
-            mockFindCapProjectRoot.mockResolvedValueOnce('CAPJava');
-            const result = await validateFioriAppProjectFolder('any/path');
-            expect(result).toEqual(t('validators.folderContainsCapApp'));
-        });
     });
 });