+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.1.33
+
+### Patch Changes
+
+-   Updated dependencies [d148718]
+    -   @sap-ux/cf-deploy-config-writer@0.1.26
+
+## 0.1.32
+
+### Patch Changes
+
+-   Updated dependencies [15222c2]
+    -   @sap-ux/cf-deploy-config-writer@0.1.25
+
+## 0.1.31
+
+### Patch Changes
+
+-   Updated dependencies [ced5edf]
+    -   @sap-ux/inquirer-common@0.6.32
+    -   @sap-ux/deploy-config-generator-shared@0.0.39
+    -   @sap-ux/cf-deploy-config-inquirer@0.2.22
+
 ## 0.1.30
 
 ### Patch Changes