+++ packages/ui5-library-sub-generator/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/ui5-library-sub-generator
 
+## 0.0.77
+
+### Patch Changes
+
+-   @sap-ux/ui5-library-inquirer@0.3.63
+
+## 0.0.76
+
+### Patch Changes
+
+-   @sap-ux/ui5-library-inquirer@0.3.62
+
+## 0.0.75
+
+### Patch Changes
+
+-   @sap-ux/ui5-library-inquirer@0.3.61
+
 ## 0.0.74
 
 ### Patch Changes